requirements.txt
asgiref==3.3.4
auto-sklearn==0.12.6
click==8.0.1
cloudpickle==1.6.0
ConfigSpace==0.4.19
contextvars==2.4
Cython==0.29.23
dask==2021.3.0
distributed==2021.3.0
Django==3.2.4
djangorestframework==3.12.4
gunicorn==20.1.0
HeapDict==1.0.1
immutables==0.15
importlib-metadata==4.5.0
joblib==1.0.1
lazy-import==0.2.2
liac-arff==2.5.0
msgpack==1.0.2
numpy==1.19.5
pandas==1.1.5
psutil==5.8.0
pynisher==0.6.4
pyparsing==2.4.7
pyrfr==0.8.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
smac==0.13.1
sortedcontainers==2.4.0
sqlparse==0.4.1
tblib==1.7.0
threadpoolctl==2.1.0
toolz==0.11.1
tornado==6.1
typing-extensions==3.10.0.0
zict==2.0.0
zipp==3.4.1
